requirements.txt
flask~=1.1.2
flask-restful
flask_sqlalchemy
flask_migrate
flask_marshmallow
marshmallow_sqlalchemy
SQLAlchemy~=1.3.22
flasgger~=0.9.5
psycopg2-binary
bcrypt
flask_mail
PyJWT